complete doc 12

+++ publications/app/Http/Controllers/PublicationController.php
@@ -13,8 +13,8 @@ class PublicationController extends Controller
     public function index()
     {
         $publications = Publication::all();
-        foreach($publications as $p) {
-            if(is_null($p->author->deleted_at)){
+        foreach ($publications as $p) {
+            if (is_null($p->author->deleted_at)) {
                 $isDeleted = false;
             } else {
                 $isDeleted = Carbon::parse($p->author->deleted_at)->isPast();
@@ -30,7 +30,7 @@ class PublicationController extends Controller
     {
         $dt = Carbon::create($publication['created_at']);
         $time = $dt->diffForHumans();
-        if(is_null($publication->author->deleted_at)){
+        if (is_null($publication->author->deleted_at)) {
             $isDeleted = false;
         } else {
             $isDeleted = Carbon::parse($publication->author->deleted_at)->isPast();
@@ -39,15 +39,16 @@ class PublicationController extends Controller
             'publication' => $publication,
             'time' => $time,
             'isdeleted' => $isDeleted
-        ]);        
+        ]);
     }
-    public function create() 
+    public function create()
     {
         return view('form', ['authors' => User::all()]);
     }
 
     public function edit(Publication $publication)
     {
+        $this->authorize('update', $publication);
         return view('form', [
             'publication' => $publication,
             'authors' => User::all()
@@ -56,6 +57,7 @@ class PublicationController extends Controller
 
     public function update(UpdatePublicationRequest $request, Publication $publication)
     {
+        $this->authorize('update', $publication);
         $data = $request->validated();
         $publication->fill($data);
         $publication->save();
@@ -63,7 +65,7 @@ class PublicationController extends Controller
     }
 
     public function store(StorePublicationRequest $request)
-    {   
+    {
         $data = $request->validated();
         $newPublication = new Publication($data);
         $newPublication->save();
@@ -74,6 +76,7 @@ class PublicationController extends Controller
 
     public function destroy(Publication $publication)
     {
+        $this->authorize('destroy', $publication);
         $publication->comment()->delete();
         $publication->delete();
         return redirect()->route('publications.index')->with('success', 'Publikacja została usunięta');
